+++ client/src/components/Form/SignUpFormTypes.ts
@@ -1,0 +1,5 @@
+import type { SignUpRequest } from '@apis/auth';
+
+export interface SignUpFormFields extends SignUpRequest {
+  passwordCheck: string;
+}
